+++ python_modules/libraries/dagster-components/dagster_components/components/dbt_project/component.py
@@ -24,11 +24,7 @@ from dagster_components.core.schema.objects import (
     OpSpecModel,
     ResolutionContext,
 )
-from dagster_components.core.schema.resolvable_from_schema import (
-    FieldResolver,
-    ResolvableModel,
-    ResolvedFrom,
-)
+from dagster_components.resolved.model import FieldResolver, ResolvableModel, ResolvedFrom
 from dagster_components.scaffoldable.decorator import scaffoldable
 from dagster_components.utils import TranslatorResolvingInfo, get_wrapped_translator_class
