+++ python_modules/libraries/dagster-components/dagster_components/components/definitions_component/component.py
@@ -11,7 +11,7 @@ from dagster_components import Component, ComponentLoadContext
 from dagster_components.components.definitions_component.scaffolder import (
     DefinitionsComponentScaffolder,
 )
-from dagster_components.core.schema.resolvable_from_schema import ResolvableModel
+from dagster_components.resolved.model import ResolvableModel
 from dagster_components.scaffoldable.decorator import scaffoldable
