+++ python_modules/libraries/dagster-components/dagster_components/core/schema/objects.py
@@ -17,7 +17,7 @@ from pydantic import BaseModel, Field
 from typing_extensions import TypeAlias
 
 from dagster_components.resolved.context import ResolutionContext
-from dagster_components.core.schema.resolvable_from_schema import (
+from dagster_components.resolved.model import (
     FieldResolver,
     ResolvableModel,
     ResolvedFrom,
