+++ python_modules/libraries/dagster-components/dagster_components/resolved/model.py
