+++ python_modules/libraries/dagster-components/dagster_components/utils.py
@@ -18,11 +18,11 @@ from dagster._core.definitions.declarative_automation.automation_condition impor
 from dagster._core.definitions.tags import build_kind_tag
 from dagster._core.errors import DagsterError
 
-from dagster_components.resolved.context import ResolutionContext
 from dagster_components.core.schema.objects import (
     AssetAttributesModel,
     resolve_asset_attributes_to_mapping,
 )
+from dagster_components.resolved.context import ResolutionContext
 
 T = TypeVar("T")
