+++ python_modules/libraries/dagster-components/dagster_components_tests/resolution_tests/custom_scope_component/component.py
@@ -5,7 +5,7 @@ from typing import Any
 from dagster import AssetSpec, AutomationCondition, Definitions
 from dagster_components import AssetAttributesModel, Component, ComponentLoadContext
 from dagster_components.core.schema.objects import ResolvedAssetAttributes
-from dagster_components.core.schema.resolvable_from_schema import ResolvableModel, ResolvedFrom
+from dagster_components.resolved.model import ResolvableModel, ResolvedFrom
 
 
 def my_custom_fn(a: str, b: str) -> str:
