+++ python_modules/libraries/dagster-components/dagster_components_tests/resolution_tests/test_field_utilites.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass
 from typing import Annotated
 
-from dagster_components.core.schema.resolvable_from_schema import (
+from dagster_components.resolved.model import (
     FieldResolver,
     ResolvedKwargs,
     get_annotation_field_resolvers,
