+++ python_modules/libraries/dagster-components/dagster_components_tests/resolution_tests/test_resolvable_from_schema.py
@@ -1,7 +1,7 @@
 from typing import Annotated
 
 from dagster_components.resolved.context import ResolutionContext
-from dagster_components.core.schema.resolvable_from_schema import (
+from dagster_components.resolved.model import (
     FieldResolver,
     ResolvableModel,
     ResolvedFrom,
