+++ python_modules/libraries/dagster-components/dagster_components_tests/unit_tests/test_spec_processing.py
@@ -2,12 +2,12 @@ from collections.abc import Sequence
 
 import pytest
 from dagster import AssetKey, AssetSpec, AutomationCondition, Definitions
-from dagster_components.resolved.context import ResolutionContext
 from dagster_components.core.schema.objects import (
     AssetAttributesModel,
     AssetPostProcessorModel,
     apply_post_processor_to_defs,
 )
+from dagster_components.resolved.context import ResolutionContext
 from pydantic import BaseModel, TypeAdapter
